requirements.txt
python-dotenv>=1.0.0,<2.0.0
langgraph>=0.1.0,<0.4.0
langchain-openai>=0.1.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
pydantic>=2.6.0,<3.0.0
typing-extensions>=4.10.0
